refactor: centralize command extraction
